requirements.txt
# ─── Hand Gesture OS Controller (Lite) ─ requirements.txt ───────────────────
#
# Install with:  pip install -r requirements.txt
#
# Python >= 3.10 required (tested on 3.12)
# ─────────────────────────────────────────────────────────────────────────────

# Computer-vision / ML  (Tasks API: HandLandmarker)
mediapipe>=0.10.14

# OpenCV – standard build
opencv-python>=4.8.0

# OS-level mouse / keyboard control
pyautogui>=0.9.54

# Modern dark-mode GUI toolkit
customtkinter>=5.2.0

# Image handling for GUI embedding
Pillow>=10.0.0

# Utility
numpy>=1.24.0
